Refactored components
