+++ ecommerce-ui/src/components/bnbRentalCard/BnbRentalContent2.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Grid, Typography } from "@mui/material/";
 import PropTypes from "prop-types";
-import AddToCart from "../AddToCart";
+import AddToCart from "./AddToCart";
 
 export default function BnbRentalContent2({
   id,
@@ -47,7 +47,6 @@ export default function BnbRentalContent2({
 BnbRentalContent2.propTypes = {
   id: PropTypes.number.isRequired,
   houseType: PropTypes.string.isRequired,
-
   location: PropTypes.shape({
     city: PropTypes.string.isRequired,
     country: PropTypes.string.isRequired,
