+++ ecommerce-ui/src/components/bnbRentalCard/BnbRentalImage.jsx
